Function clicolors_control::configure_terminal

pub fn configure_terminal() -> bool

Configures the terminal for ANSI color support.

This is not needed on UNIX systems and normally automatically happens on windows the first time colors_enabled() is called. This automatic behavior however can be disabled by removing the terminal_autoconfig feature flag.

When this function is called and the terminal was reconfigured, changes from set_colors_enabled are reverted.

It returns true if the terminal supports colors after configuration or false if not.